[#635] Removed non working service, using the backup one for all devices
